Set aws-base Version to latest but add Revision
Since we will always be using 'latest' version of aws-base,
we need to set it accordingly.

However, sometimes we may need to force-rebuild the image
without other changes to Dockerfile or entrypoint.sh
(e.g. to update its packages).

As such, we introduce a Revision number that will instruct
build.sh to re-build aws-base based on a new hash.